+++ lib/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(Coro)
 add_subdirectory(WAMI)
 add_subdirectory(wasmstack)
 add_subdirectory(DCont)
