+++ test/integration/coro/coro-oneshot-trace-runtime.mlir
