+++ wasm-opt/wasm-opt.cpp
@@ -6,6 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "Coro/CoroPasses.h"
 #include "DCont/DContDialect.h"
 #include "WAMI/WAMIDialect.h"
 #include "WAMI/WAMIPasses.h"
@@ -29,6 +30,7 @@
 
 int main(int argc, char **argv) {
   mlir::registerAllPasses();
+  mlir::coro::registerPasses();
   mlir::wami::registerPasses();
   mlir::wasmstack::registerPasses();
